Use preproc from mzm

+++ Makefile
@@ -41,6 +41,7 @@ PREPROC = tools/preproc/preproc
 ASFLAGS = -mcpu=arm7tdmi
 CFLAGS = -Werror -O2 -mthumb-interwork -fhex-asm
 CPPFLAGS = -nostdinc -Iinclude/
+PREPROCFLAGS = charmap.txt
 
 # Objects
 CSRC = $(wildcard src/**.c) $(wildcard src/**/**.c) $(wildcard src/**/**/**.c)
@@ -75,20 +76,29 @@ diff: $(DUMPS)
 
 .PHONY: clean
 clean:
-	$(MSG) RM $(TARGET)
-	$Q$(RM) $(TARGET) $(ELF) $(MAP)
+	$(MSG) RM roms
+# Delete every gba file that doesn't end with baserom
+	$Qfind -type f -name "*.gba" -a ! -name "*baserom.gba" -delete
+	$(MSG) RM elf
+	$Q$(RM) *.elf
+	$(MSG) RM map
+	$Q$(RM) *.map
+
 	$(MSG) RM \*.dump
 	$Q$(RM) $(DUMPS)
 	$(MSG) RM \*.o
 	$Q$(RM) $(OBJ)
 	$(MSG) RM data/*.s
-	$Q$(RM) $(DATA)
 	$(MSG) RM src/\*\*/\*.s
 	$Q$(RM) $(CSRC:.c=.s)
 	$(MSG) RM $(GBAFIX)
 	$Q$(RM) $(GBAFIX)
+ifeq ($(DATA),1)
 	$(MSG) RM data/
-	$Q$(RM) -r data	
+	$Q$(RM) -r data
+endif
+	$(MSG) RM linker.ld.pp
+	$Q$(RM) linker.ld.pp
 
 .PHONY: help
 help:
@@ -123,7 +133,7 @@ $(ELF) $(MAP): $(OBJ) linker.ld
 
 %.s: %.c
 	$(MSG) CC $@
-	$Q$(PREPROC) $< | $(CPP) $(CPPFLAGS) | $(CC) -o $@ $(CFLAGS) && printf '\t.align 2, 0 @ dont insert nops\n' >> $@
+	$Q$(PREPROC) $< $(PREPROCFLAGS) | $(CPP) $(CPPFLAGS) | $(CC) -o $@ $(CFLAGS) && printf '\t.align 2, 0 @ dont insert nops\n' >> $@
 
 src/sram/%.s: CFLAGS = -O1 -mthumb-interwork -fhex-asm
 src/sram/%.s: src/sram/%.c
